page_list: use a dict to keep track of pages.
Uses an additional dictionary but walking the list to test for
an existing URL is far too slow. This uses more mem, but it's
simpler and faster.

We can also ensure we don't duplicate with our SQL, but in regards
to crawling and client-side tracking of URLs, this is fine, it
does the job...