config: use INI values for now.
Need to think of some good way to cache and pass credentials
to the docker container instances.

Perhaps it's easier to ensure they use a local config file and
anything on the raw host uses encryption/password vault.

Having to query a remote server for credentials seems insane, but,
it's seemingly a common practice.