form: robotlog form.
Learning a lot here. Whether I can put it into words is another
thing. We have, at least, the start of a form.

Aim to keep this program simple. In procedural code I was able
to do much more complicated stuff, but then, nobody could read it.

Perhaps...in time...the same with Symfony. For now, keep it simple.